Refactor JSON Serializer for logs
