+++ src/callcenter_billing/settings.py
@@ -11,7 +11,6 @@ https://docs.djangoproject.com/en/2.0/ref/settings/
 """
 
 import os
-import json_log_formatter
 import django_heroku
 
 
@@ -149,7 +148,7 @@ LOGGING = {
             'format': '%(levelname)s %(message)s'
         },
         'json': {
-            '()': 'callcenter_billing.settings.CustomisedJSONFormatter',
+            '()': 'callcenter_billing.serializers.CustomisedJSONFormatter',
         }
     },
     'handlers': {
@@ -172,25 +171,5 @@ LOGGING = {
         },
     },
 }
-
-
-class CustomisedJSONFormatter(json_log_formatter.JSONFormatter):
-
-    def mutate_json_record(self, json_record):
-        from datetime import datetime, time, timedelta
-
-        for attr_name in json_record:
-            attr = json_record[attr_name]
-            if isinstance(attr, datetime):
-                json_record[attr_name] = attr.isoformat()
-            elif isinstance(attr, time):
-                json_record[attr_name] = attr.isoformat()
-            elif isinstance(attr, timedelta):
-                json_record[attr_name] = str(attr)
-        return json_record
-
-    def json_record(self, message, extra, record):
-        extra = super().json_record(message, extra, record)
-        return extra
 
 django_heroku.settings(locals())
